+++ tools/tf-hcl-graph/testdata/for_each_empty/main.tf
@@ -1,0 +1,6 @@
+resource "azurerm_storage_account" "empty_example" {
+  for_each = {}
+
+  name     = "st${each.key}"
+  location = each.value
+}