+++ tools/tf-hcl-graph/testdata/for_each_set/main.tf
@@ -1,0 +1,6 @@
+resource "azurerm_storage_account" "set_example" {
+  for_each = toset(["web", "db"])
+
+  name     = "st${each.key}"
+  location = each.value
+}