+++ tools/tf-hcl-graph/testdata/module_dynamic/child/main.tf
@@ -1,0 +1,8 @@
+variable "location" {
+  type = string
+}
+
+resource "azurerm_storage_account" "this" {
+  name     = "st"
+  location = var.location
+}