+++ tools/tf-hcl-graph/testdata/module_dynamic/main.tf
@@ -1,0 +1,11 @@
+variable "envs" {
+  type    = map(string)
+  default = {}
+}
+
+module "storage" {
+  source   = "./child"
+  for_each = var.envs
+
+  location = each.value
+}